feat: implement AttendanceMapper
